+++ CHANGELOG.md
@@ -2,6 +2,8 @@ In Development
 --------------
 - Deployment:
     - Configure systemd's journaling to only retain the last 60 days of logs
+- Dependencies:
+    - Update wheel-filename to 2.1+
 
 v2025.11.21
 -----------